+++ CHANGES.rst
@@ -22,6 +22,9 @@ To be released.
 - Since returning a disallowed value which does not match to the return type
   is the fault the server-side made, the HTTP status code for the case became
   ``500 Internal Server Error`` instead of ``400 Bad Request``.
+- For the mistakes returning ``None`` from a method having non-null return type,
+  now it became to show a more readable and debug-friendly message with a proper
+  response instead of uncaught Python exception.
 - ``WsgiApp.url_map`` attribute was gone.
 - ``/ping/`` resource was gone.